Some refactoring to get dsptools working with these examples
