+++ generators/chipyard/src/main/scala/example/dsptools/DspBlocks.scala
@@ -1,4 +1,4 @@
-package example
+package chipyard.example
 
 import chisel3._
 import chisel3.util._
